requirements.txt
Flask
Gunicorn
PyJWT
flask-cors
